# docs/openapi.generated.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "제대로 API",
    "description": "현역 장병용 피부 셀프케어 PWA API",
    "version": "v1"
  },
  "servers": [
    {
      "url": "https://api.flourishing.o-r.kr",
      "description": "운영"
    },
    {
      "url": "http://localhost:8080",
      "description": "로컬"
    }
  ],
  "security": [
    {
      "cookieAuth": []
    }
  ],
...
  "paths": {
    "/v1/skin-reports/{reportId}/follow-up": {
      "get": {
        "tags": [
          "FollowUps"
        ],
        "summary": "다음 날 경과 조회",
        "operationId": "getFollowUp",
        "parameters": [
          {
            "name": "reportId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/FollowUpResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "FollowUps"
        ],
        "summary": "다음 날 경과 저장",
        "operationId": "saveFollowUp",
        "parameters": [
          {
            "name": "reportId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/ClinicianFollowUpRequest"
                  },
                  {
                    "$ref": "#/components/schemas/SelfCareFollowUpRequest"
                  }
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "경과 저장 및 보고 COMPLETED 전환",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/FollowUpResponse"
                }
              }
            }
          },
          "200": {
            "description": "같은 경과가 이미 저장되어 있음",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/FollowUpResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/push-subscriptions": {
      "post": {
        "tags": [
          "Notifications"
        ],
        "summary": "Push 구독 등록",
        "operationId": "register_1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterPushSubscriptionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "새 구독 등록",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PushSubscriptionResponse"
                }
              }
            }
          },
          "200": {
            "description": "같은 endpoint 의 기존 구독 갱신",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PushSubscriptionResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/push-subscriptions/{subscriptionId}": {
      "delete": {
        "tags": [
          "Notifications"
        ],
        "summary": "Push 구독 해제",
        "operationId": "unregister",
        "parameters": [
          {
            "name": "subscriptionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "구독 해제 완료"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ClinicianFollowUpRequest": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SaveFollowUpRequest"
          },
          {
            "type": "object",
            "properties": {
              "kind": {
                "type": "string",
                "enum": [
                  "SELF_CARE",
                  "CLINICIAN_CHECK"
                ]
              },
              "skinChange": {
                "type": "string",
                "enum": [
                  "IMPROVED",
                  "SIMILAR",
                  "WORSENED"
                ]
              },
              "actionCompletion": {
                "type": "string",
                "enum": [
                  "MOSTLY_DONE",
                  "PARTLY_DONE",
                  "NOT_DONE"
                ]
              },
              "clinicianCheckStatus": {
                "type": "string",
                "enum": [
                  "CHECKED",
                  "NOT_YET",
                  "PREFER_NOT_TO_RECORD"
                ]
              }
            }
          }
        ],
        "required": [
          "actionCompletion",
          "clinicianCheckStatus",
          "kind",
          "skinChange"
        ]
      },
      "SaveFollowUpRequest": {
        "type": "object",
        "discriminator": {
          "propertyName": "kind"
        },
        "properties": {
          "kind": {
            "type": "string"
          }
        },
        "required": [
          "kind"
        ]
      },
      "SelfCareFollowUpRequest": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SaveFollowUpRequest"
          },
          {
            "type": "object",
            "properties": {
              "kind": {
                "type": "string",
                "enum": [
                  "SELF_CARE",
                  "CLINICIAN_CHECK"
                ]
              },
              "skinChange": {
                "type": "string",
                "enum": [
                  "IMPROVED",
                  "SIMILAR",
                  "WORSENED"
                ]
              },
              "actionCompletion": {
                "type": "string",
                "enum": [
                  "MOSTLY_DONE",
                  "PARTLY_DONE",
                  "NOT_DONE"
                ]
              }
            }
          }
        ],
        "required": [
          "actionCompletion",
          "kind",
          "skinChange"
        ]
      },
      "FollowUpResponse": {
        "type": "object",
        "properties": {
          "reportId": {
            "type": "string",
            "format": "uuid"
          },
          "kind": {
            "type": "string",
            "enum": [
              "SELF_CARE",
              "CLINICIAN_CHECK"
            ]
          },
          "skinChange": {
            "type": "string",
            "enum": [
              "IMPROVED",
              "SIMILAR",
              "WORSENED"
            ]
          },
          "actionCompletion": {
            "type": "string",
            "enum": [
              "MOSTLY_DONE",
              "PARTLY_DONE",
              "NOT_DONE"
            ]
          },
          "clinicianCheckStatus": {
            "type": "string",
            "enum": [
              "CHECKED",
              "NOT_YET",
              "PREFER_NOT_TO_RECORD"
            ]
          },
          "submittedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "PushSubscriptionKeysRequest": {
        "type": "object",
        "properties": {
          "p256dh": {
            "type": "string",
            "maxLength": 512,
            "minLength": 20
          },
          "auth": {
            "type": "string",
            "maxLength": 256,
            "minLength": 8
          }
        }
      },
      "RegisterPushSubscriptionRequest": {
        "type": "object",
        "properties": {
          "endpoint": {
            "type": "string",
            "maxLength": 2048,
            "minLength": 0,
            "pattern": "^https://[^\\s]+$"
          },
          "expirationTime": {
            "type": "string"
          },
          "keys": {
            "$ref": "#/components/schemas/PushSubscriptionKeysRequest"
          },
          "userAgent": {
            "type": "string",
            "maxLength": 512,
            "minLength": 0
          }
        },
        "required": [
          "keys"
        ]
      },
      "PushSubscriptionResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "endpointFingerprint": {
            "type": "string"
          },
          "active": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    },
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "description": "Secure, HttpOnly, SameSite=Lax 세션 쿠키",
        "name": "__Host-session",
        "in": "cookie"
      },
      "csrfToken": {
        "type": "apiKey",
        "description": "현재 세션 응답에서 받은 CSRF 토큰",
        "name": "X-CSRF-Token",
        "in": "header"
      }
    }
  }
}